# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.FileResponse:
    properties:
      id:
        type: integer
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
      thumbnail:
        type: string
      url:
        type: string
    type: object
...
  service.FrontendAuthor:
    properties:
      avatar:
        $ref: '#/definitions/service.FileResponse'
      name:
        type: string
    type: object
  service.FrontendMoment:
    properties:
      author:
        $ref: '#/definitions/service.FrontendAuthor'
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/service.FrontendPhoto'
        type: array
      isPublic:
        type: boolean
      likes:
        type: integer
    type: object
  service.FrontendPhoto:
    properties:
      file:
        $ref: '#/definitions/service.FileResponse'
      id:
        type: integer
      momentId:
        type: integer
    type: object
...
  service.MomentCreateRequest:
    properties:
      content:
        type: string
      imageIds:
        items:
          type: integer
        type: array
      isPublic:
        type: boolean
      userId:
        type: integer
    required:
    - content
    - userId
    type: object
  service.MomentLikeResponse:
    properties:
      likes:
        type: integer
    type: object
  service.MomentListResponse:
    properties:
      moments:
        items:
          $ref: '#/definitions/service.FrontendMoment'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  service.MomentPublicRequest:
    properties:
      isPublic:
        type: boolean
    type: object
  service.MomentUpdateRequest:
    properties:
      content:
        type: string
      imageIds:
        items:
          type: integer
        type: array
      isPublic:
        type: boolean
    type: object
  service.PlaceCreateRequest:
    properties:
      date:
        type: string
      description:
        type: string
      image:
        $ref: '#/definitions/service.PlaceImage'
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      name:
        type: string
    required:
    - date
    - latitude
    - longitude
    - name
    type: object
  service.PlaceImage:
    properties:
      file:
        $ref: '#/definitions/service.FileResponse'
      id:
        type: integer
      placeId:
        type: integer
    type: object
  service.PlaceListResponse:
    properties:
      page:
        type: integer
      places:
        items:
          $ref: '#/definitions/service.PlaceResponse'
        type: array
      size:
        type: integer
      total:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
    type: object
  service.PlaceResponse:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        $ref: '#/definitions/service.PlaceImage'
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  service.PlaceUpdateRequest:
    properties:
      date:
        type: string
      description:
        type: string
      image:
        $ref: '#/definitions/service.PlaceImage'
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/v1/places:
    get:
      consumes:
      - application/json
      description: 获取所有地点，保持分页数据结构
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.PlaceListResponse'
              type: object
      summary: 获取所有地点
      tags:
      - places
    post:
      consumes:
      - application/json
      description: 创建一个新的地点
      parameters:
      - description: 地点信息
        in: body
        name: place
        required: true
        schema:
          $ref: '#/definitions/service.PlaceCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.PlaceResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建地点
      tags:
      - places
  /api/v1/places/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的地点
      parameters:
      - description: 地点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 删除地点
      tags:
      - places
    put:
      consumes:
      - application/json
      description: 更新地点的信息
      parameters:
      - description: 地点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 地点信息
        in: body
        name: place
        required: true
        schema:
          $ref: '#/definitions/service.PlaceUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.PlaceResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新地点
      tags:
      - places
...
  /moments:
    get:
      description: 分页获取动态列表
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MomentListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取动态列表
      tags:
      - moments
    post:
      consumes:
      - application/json
      description: 创建一个新的动态
      parameters:
      - description: 动态信息
        in: body
        name: moment
        required: true
        schema:
          $ref: '#/definitions/service.MomentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.FrontendMoment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 创建动态
      tags:
      - moments
  /moments/{id}:
    delete:
      description: 删除指定ID的动态
      parameters:
      - description: 动态ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data: {}
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 删除动态
      tags:
      - moments
    put:
      consumes:
      - application/json
      description: 更新动态信息
      parameters:
      - description: 动态ID
        in: path
        name: id
        required: true
        type: string
      - description: 动态信息
        in: body
        name: moment
        required: true
        schema:
          $ref: '#/definitions/service.MomentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.FrontendMoment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 更新动态
      tags:
      - moments
  /moments/{id}/like:
    post:
      description: 为指定ID的动态点赞
      parameters:
      - description: 动态ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MomentLikeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 点赞动态
      tags:
      - moments
  /moments/{id}/public:
    put:
      consumes:
      - application/json
      description: 更新动态的公开状态
      parameters:
      - description: 动态ID
        in: path
        name: id
        required: true
        type: string
      - description: 公开状态
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/service.MomentPublicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.FrontendMoment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 更新动态公开状态
      tags:
      - moments
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      admin: Grants read and write access to administrative information
      read: Grants read access
      write: Grants write access
    tokenUrl: /api/v1/user/token
    type: oauth2
swagger: "2.0"